+++ bin/c-kc-apply.js
@@ -14,6 +14,7 @@ const {
 const { formatProjectPrefix } = require('./lib/c-project');
 const {
     ensureServiceFilesExist,
+    removeCompiledSecrets,
     renderServicesTemplates,
     setLocalsForServices,
 } = require('./lib/c-kc');
@@ -100,11 +101,11 @@ return Promise.all([loadConfig(), loadState()])
                     return reject(e);
                 }
 
-                return resolve([services, path]);
+                return resolve([services, path, state]);
             })
     )
     .then(
-        ([services, path]) =>
+        ([services, path, state]) =>
             new Promise(async (resolve, reject) => {
                 try {
                     await ensureServiceFilesExist(path, services);
@@ -112,11 +113,11 @@ return Promise.all([loadConfig(), loadState()])
                     reject(e);
                 }
 
-                return resolve([services, path]);
+                return resolve([services, path, state]);
             })
     )
     .then(
-        ([services, path]) =>
+        ([services, path, state]) =>
             new Promise(async (resolve, reject) => {
                 try {
                     await renderServicesTemplates(path, services);
@@ -124,11 +125,11 @@ return Promise.all([loadConfig(), loadState()])
                     reject(e);
                 }
 
-                return resolve([services, path]);
+                return resolve([services, path, state]);
             })
     )
     .then(
-        ([services, path]) =>
+        ([services, path, state]) =>
             new Promise((resolve) => {
                 services.forEach((service) => {
                     exec(
@@ -141,7 +142,11 @@ return Promise.all([loadConfig(), loadState()])
                     );
                 });
 
-                return resolve();
+                return removeCompiledSecrets({
+                    env: state.env,
+                    path,
+                    services,
+                }).then(resolve);
             })
     )
     .catch((err) => {